requirements.txt
pandas>=2.0,<3.0
requests>=2.28,<3.0
pytrends>=4.9,<5.0
beautifulsoup4>=4.12,<5.0
openpyxl>=3.1,<4.0
numpy>=1.24,<3.0
python-dotenv>=1.0,<2.0
streamlit>=1.30,<2.0
plotly>=5.18,<6.0
openai>=1.0,<3.0
